requirements.txt
google-generativeai
python-pptx
pillow
numpy